requirements.txt
torch
hydra-core==1.1.2
tensorboard